Sales Data from the Contonso Dataset

This dataset contains information about sales orders, including order details, pricing, and customer data from the Contonso dataset. It provides insights into the transactions that have occurred, including order dates, delivery dates, customer and store information, as well as product details.

Usage

sales

Format

A data frame with sales columns:

order_key

double Unique identifier for each order.

line_number

double Line number within the order (for multi-line orders).

order_date

Date Date when the order was placed.

delivery_date

Date Date when the order was delivered.

customer_key

double Unique identifier for the customer who placed the order.

store_key

double Unique identifier for the store where the order was placed.

product_key

double Unique identifier for the product in the order.

quantity

double The quantity of the product ordered.

unit_price

double The price per unit of the product.

net_price

double The total net price for the product, considering any discounts.

unit_cost

double The cost per unit of the product.

currency_code

character The currency code used for the transaction (e.g., USD, EUR).

exchange_rate

double The exchange rate applied to the currency, if applicable.

gross_revenue

double A product's unit_price multiplied by quantity.

net_revenue

double A product's net_price multiplied by quantity.

unit_discount

double A product's unit_price minute net_price.

discounts

double A product's unit_discount multiplied by quantity.

cogs

double A product's unit_cost multiplied by quantity.

margin

double A product's net_revenue minus cogs.

unit_margin

double A product margin divided by quantity.

Source

https://github.com/sql-bi/Contoso-Data-Generator-V2-Data/releases/tag/ready-to-use-data